DBZ-2994 Moved Scn to parent package
Since the Scn class is used by the snapshot process, it makes sense to move
this class up by one package since it can now be leveraged by both adapters
as needed.
