+++ debezium-connector-oracle/src/main/java/io/debezium/connector/oracle/logminer/NeverHistoryRecorder.java
@@ -7,6 +7,7 @@ package io.debezium.connector.oracle.logminer;
 
 import java.sql.Timestamp;
 
+import io.debezium.connector.oracle.Scn;
 import io.debezium.jdbc.JdbcConfiguration;
 
 /**
