+++ debezium-connector-oracle/src/main/java/io/debezium/connector/oracle/logminer/RowMapper.java
@@ -14,6 +14,7 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import io.debezium.DebeziumException;
+import io.debezium.connector.oracle.Scn;
 import io.debezium.relational.TableId;
 import io.debezium.util.HexConverter;
